Remove remaining svn $Id$ tags (examples/ dir)
This removes the $Id$ tags from the info output in the examples files.
Previously, the id information was printed as part of running a given
example.  This information isn't really helpful anymore in a project using
Git as a version control system, hence the informative messages have been
reduced to merely say which file is being run (which is largely what the
output previously did).