Use use_ok() in basic test
The test `01basic.t` was `use`-ing the `Device::Gsm` module and checking if
it had been loaded via an `END` block, which made the code a bit complex.
This test was written in 2002 and it's likely that the `use_ok` function
didn't exist back then.  Using the new function simplifies the test
significantly.