Update link to pduconv
Actually, the pduconv project seems to no longer exist.  I even searched for
Mats Engstrom (the original developer) and he now seems to work in Bangkok
and various links of his lead to unconfigured or hacked webservers.  The
trail to the actual code seems to peter out at the updated links in this
commit.