Use Device::Gsm version for D::G::Charset VERSION
The version set in D::G::Charset used the Subversion `$Revision$` tag to
determine the version number.  Since the project is now hosted in Git, this
no longer makes sense.  The best thing to do is thus to tie D::G::Charset's
version number to that of Device::Gsm itself.